files(5)/requirements.txt
# Installa queste dipendenze nel sistema (non nel venv pix2tex)
# pip install -r requirements.txt

python-pptx>=0.6.21
Pillow>=10.0.0
numpy>=1.24.0
lxml>=4.9.0
